requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-cpp-python==0.2.90
python-dotenv==1.0.0
